requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.8
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.1.0
pkg-resources==0.0.0
pymisp==2.4.101
python-dateutil==2.7.5
PyYAML==3.13
requests==2.21.0
sigmatools==0.7.1
six==1.12.0
urllib3==1.24.1
